// src/app/features/trade-pnl-query/models/profit.model.ts
export interface Profit {
  mtime: string;
  dseq: string;
  ttypename: string;
  bstypename: string;
  stock: string;
  stocknm: string;
  qty: string;
  price: string;
  mamt: string;
  cramt: string;
  dnamt: string;
  gtamt: string;
  fee: string;
  tax: string;
  interest: string;
  dbfee: string;
  netamt: string;
  testpro: string;
  ts_ratio: string;
  cost: string;
  currnm: string;
}